@@ -4,8 +4,7 @@ version: 1.0.0
 homepage: https://github.com/charlee/simple_animated_icon
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
